refactor: MetricView 개별 파일로 분리
